Refactorise l'accès rapide du header de l'app
