--- apps/app/src/app/Layout/Header/MenuUtilisateur.css
@@ -1,23 +1,0 @@
-.user-menu .fr-nav__link:hover {
-  --underline-hover-width: 0;
-}
-
-.user-menu .fr-nav__link {
-  --underline-img: linear-gradient(0deg, #ccc, #ccc);
-}
-
-.user-menu .fr-nav__link[aria-current]::before {
-  left: 0.5rem;
-}
-
-.user-menu i::before {
-  --icon-size: 1rem;
-}
-
-button.user-menu[aria-expanded='true'] {
-  color: var(--text-active-blue-france);
-  background-color: var(--background-open-blue-france);
-  --idle: transparent;
-  --hover: var(--background-open-blue-france-hover);
-  --active: var(--background-open-blue-france-active);
-}
